requirements.txt
anthropic
openai
google-generativeai
PyPDF2
pdfplumber
python-docx
python-dotenv
openpyxl
